# fix(grasshopper): optimize `CreateSpeckleProperties` for large data trees